performance oriented spring data jpa (#7)
This pull request was based on a talk on performance tips for Spring Data JPA. The following things were considered in the talk.

Instead of blindly guessing performance-related problems, it is best to set up proper observability tools for monitoring application performance.
Database connections are managed through connection pools like HikariCP. The number of connections in the connection pool should be minimal and optimally around the magnitude of the number of cores to prevent overhead from context switching between database connection threads.
A major performance overhead is the database connection not closing when it should be closed. If the OSIV feature is turned on (which is the default), the database connection is opened until the web view stage - it is best to turn this feature off. If auto-commits are on for database transactions (which is the default), then the connection is opened even when its not needed - it is best to turn this feature off. If the transaction within the method is complex, we can control the transaction programatically.
Another problem was around fetching entities. Eager fetching should be avoided, and if associations should be fetched, fetch joins or entity graphs should be used.
Projections should be utilized to fetch only the needed columns. If there are way to many projections, we can use dynamic projections.
Add a version field or implement the Persistable interface for entities. This is because when an entity is saved, JPA doesn't know if the entity is new or not, therefore JPA fetches an entity before it gets persisted.
Saving entities can be an overhead if there are a lot of fields but we update a single column. This can be avoided by adding the dynamic update annotation to the entity. JPA will track only the dirty columns.
Sometimes JPA might not be a right fit for applications, and using raw SQL could do the job.